@@ -1,9 +1,10 @@
 package com.paginate.recycler;
 
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 
 import com.paginate.Paginate;
 
@@ -15,12 +16,14 @@ public final class RecyclerPaginate extends Paginate {
     private WrapperAdapter wrapperAdapter;
     private WrapperSpanSizeLookup wrapperSpanSizeLookup;
 
-    RecyclerPaginate(RecyclerView recyclerView,
-                     Paginate.Callbacks callbacks,
-                     int loadingTriggerThreshold,
-                     boolean addLoadingListItem,
-                     LoadingListItemCreator loadingListItemCreator,
-                     LoadingListItemSpanLookup loadingListItemSpanLookup) {
+    RecyclerPaginate(
+        RecyclerView recyclerView,
+        Paginate.Callbacks callbacks,
+        int loadingTriggerThreshold,
+        boolean addLoadingListItem,
+        LoadingListItemCreator loadingListItemCreator,
+        LoadingListItemSpanLookup loadingListItemSpanLookup
+    ) {
         this.recyclerView = recyclerView;
         this.callbacks = callbacks;
         this.loadingTriggerThreshold = loadingTriggerThreshold;
@@ -31,6 +34,7 @@ public final class RecyclerPaginate extends Paginate {
         if (addLoadingListItem) {
             // Wrap existing adapter with new adapter that will add loading row
             RecyclerView.Adapter adapter = recyclerView.getAdapter();
+            assert adapter != null;
             wrapperAdapter = new WrapperAdapter(adapter, loadingListItemCreator);
             adapter.registerAdapterDataObserver(mDataObserver);
             recyclerView.setAdapter(wrapperAdapter);
@@ -38,9 +42,10 @@ public final class RecyclerPaginate extends Paginate {
             // For GridLayoutManager use separate/customisable span lookup for loading row
             if (recyclerView.getLayoutManager() instanceof GridLayoutManager) {
                 wrapperSpanSizeLookup = new WrapperSpanSizeLookup(
-                        ((GridLayoutManager) recyclerView.getLayoutManager()).getSpanSizeLookup(),
-                        loadingListItemSpanLookup,
-                        wrapperAdapter);
+                    ((GridLayoutManager) recyclerView.getLayoutManager()).getSpanSizeLookup(),
+                    loadingListItemSpanLookup,
+                    wrapperAdapter
+                );
                 ((GridLayoutManager) recyclerView.getLayoutManager()).setSpanSizeLookup(wrapperSpanSizeLookup);
             }
         }
@@ -73,7 +78,7 @@ public final class RecyclerPaginate extends Paginate {
         }
     }
 
-    void checkEndOffset() {
+    private void checkEndOffset() {
         int visibleItemCount = recyclerView.getChildCount();
         int totalItemCount = recyclerView.getLayoutManager().getItemCount();
 
@@ -92,8 +97,7 @@ public final class RecyclerPaginate extends Paginate {
         }
 
         // Check if end of the list is reached (counting threshold) or if there is no items at all
-        if ((totalItemCount - visibleItemCount) <= (firstVisibleItemPosition + loadingTriggerThreshold)
-                || totalItemCount == 0) {
+        if ((totalItemCount - visibleItemCount) <= (firstVisibleItemPosition + loadingTriggerThreshold) || totalItemCount == 0) {
             // Call load more only if loading is not currently in progress and if there is more items to load
             if (!callbacks.isLoading() && !callbacks.hasLoadedAllItems()) {
                 callbacks.onLoadMore();
@@ -108,7 +112,7 @@ public final class RecyclerPaginate extends Paginate {
 
     private final RecyclerView.OnScrollListener mOnScrollListener = new RecyclerView.OnScrollListener() {
         @Override
-        public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
+        public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
             checkEndOffset(); // Each time when list is scrolled check if end of the list is reached
         }
     };
@@ -167,11 +171,11 @@ public final class RecyclerPaginate extends Paginate {
         }
 
         /**
-         * Set the offset from the end of the list at which the load more event needs to be triggered. Default offset
-         * if 5.
+         * Set the offset from the end of the list at which the load more event needs to be triggered.
+         * Default offset if 5.
          *
          * @param threshold number of items from the end of the list.
-         * @return {@link com.paginate.recycler.RecyclerPaginate.Builder}
+         * @return {@link RecyclerPaginate.Builder}
          */
         public Builder setLoadingTriggerThreshold(int threshold) {
             this.loadingTriggerThreshold = threshold;
@@ -180,14 +184,14 @@ public final class RecyclerPaginate extends Paginate {
 
         /**
          * Setup loading row. If loading row is used original adapter set on RecyclerView will be wrapped with
-         * internal adapter that will add loading row as the last item in the list. Paginate will observer the
+         * internal adapter that will add loading row as the last item in the list. Paginate will observe the
          * changes upon original adapter and remove loading row if there is no more data to load. By default loading
          * row will be added.
          *
          * @param addLoadingListItem true if loading row needs to be added, false otherwise.
-         * @return {@link com.paginate.recycler.RecyclerPaginate.Builder}
-         * @see {@link com.paginate.Paginate.Callbacks#hasLoadedAllItems()}
-         * @see {@link com.paginate.recycler.RecyclerPaginate.Builder#setLoadingListItemCreator(LoadingListItemCreator)}
+         * @return {@link RecyclerPaginate.Builder}
+         * @see Paginate.Callbacks#hasLoadedAllItems()
+         * @see RecyclerPaginate.Builder#setLoadingListItemCreator(LoadingListItemCreator)
          */
         public Builder addLoadingListItem(boolean addLoadingListItem) {
             this.addLoadingListItem = addLoadingListItem;
@@ -198,7 +202,7 @@ public final class RecyclerPaginate extends Paginate {
          * Set custom loading list item creator. If no creator is set default one will be used.
          *
          * @param creator Creator that will ne called for inflating and binding loading list item.
-         * @return {@link com.paginate.recycler.RecyclerPaginate.Builder}
+         * @return {@link RecyclerPaginate.Builder}
          */
         public Builder setLoadingListItemCreator(LoadingListItemCreator creator) {
             this.loadingListItemCreator = creator;
@@ -209,8 +213,8 @@ public final class RecyclerPaginate extends Paginate {
          * Set custom SpanSizeLookup for loading list item. Use this when {@link GridLayoutManager} is used and
          * loading list item needs to have custom span. Full span of {@link GridLayoutManager} is used by default.
          *
-         * @param loadingListItemSpanLookup LoadingListItemSpanLookup that will be called for loading list item span.
-         * @return {@link com.paginate.recycler.RecyclerPaginate.Builder}
+         * @param loadingListItemSpanLookup {@link LoadingListItemSpanLookup} that will be called for getting list item span.
+         * @return {@link RecyclerPaginate.Builder}
          */
         public Builder setLoadingListItemSpanSizeLookup(LoadingListItemSpanLookup loadingListItemSpanLookup) {
             this.loadingListItemSpanLookup = loadingListItemSpanLookup;
@@ -226,6 +230,7 @@ public final class RecyclerPaginate extends Paginate {
             if (recyclerView.getAdapter() == null) {
                 throw new IllegalStateException("Adapter needs to be set!");
             }
+
             if (recyclerView.getLayoutManager() == null) {
                 throw new IllegalStateException("LayoutManager needs to be set on the RecyclerView");
             }
@@ -238,9 +243,8 @@ public final class RecyclerPaginate extends Paginate {
                 loadingListItemSpanLookup = new DefaultLoadingListItemSpanLookup(recyclerView.getLayoutManager());
             }
 
-            return new RecyclerPaginate(recyclerView, callbacks, loadingTriggerThreshold, addLoadingListItem,
-                    loadingListItemCreator, loadingListItemSpanLookup);
+            return new RecyclerPaginate(recyclerView, callbacks, loadingTriggerThreshold,
+                addLoadingListItem, loadingListItemCreator, loadingListItemSpanLookup);
         }
     }
-
 }
