@@ -10,6 +10,7 @@ import com.example.paginate.R;
 import com.example.paginate.data.Person;
 
 import java.util.List;
+import java.util.Locale;
 
 public class PersonAdapter extends BindableAdapter<Person> {
 
@@ -46,7 +47,9 @@ public class PersonAdapter extends BindableAdapter<Person> {
     @Override
     public void bindView(Person person, int position, View view) {
         PersonVH holder = (PersonVH) view.getTag();
-        holder.tvFullName.setText(String.format("%s %s, %d", person.getFirstName(), person.getLastName(), person.getAge()));
+        holder.tvFullName.setText(
+            String.format(Locale.getDefault(), "%s %s, %d", person.getFirstName(), person.getLastName(), person.getAge())
+        );
     }
 
     public void add(List<Person> data) {
