@@ -1,16 +1,19 @@
 package com.example.paginate.adapter;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.example.paginate.R;
 import com.example.paginate.data.Person;
 
 import java.util.List;
+import java.util.Locale;
 
 public class RecyclerPersonAdapter extends RecyclerView.Adapter<RecyclerPersonAdapter.PersonVH> implements RecyclerOnItemClickListener {
 
@@ -20,6 +23,7 @@ public class RecyclerPersonAdapter extends RecyclerView.Adapter<RecyclerPersonAd
         this.data = data;
     }
 
+    @NonNull
     @Override
     public PersonVH onCreateViewHolder(ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.person_list_item, parent, false);
@@ -29,7 +33,9 @@ public class RecyclerPersonAdapter extends RecyclerView.Adapter<RecyclerPersonAd
     @Override
     public void onBindViewHolder(PersonVH holder, final int position) {
         Person person = data.get(position);
-        holder.tvFullName.setText(String.format("%s %s, %d", person.getFirstName(), person.getLastName(), person.getAge()));
+        holder.tvFullName.setText(
+            String.format(Locale.getDefault(), "%s %s, %d", person.getFirstName(), person.getLastName(), person.getAge())
+        );
     }
 
     @Override
