@@ -1,17 +1,20 @@
 package com.paginate.recycler;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.paginate.R;
 
-/** RecyclerView creator that will be called to create and bind loading list item */
+/**
+ * RecyclerView creator that will be called to create and bind loading list item.
+ */
 public interface LoadingListItemCreator {
 
     /**
-     * Create new loading list item {@link android.support.v7.widget.RecyclerView.ViewHolder}.
+     * Create new loading list item {@link RecyclerView.ViewHolder}.
      *
      * @param parent   parent ViewGroup.
      * @param viewType type of the loading list item.
@@ -30,7 +33,8 @@ public interface LoadingListItemCreator {
     LoadingListItemCreator DEFAULT = new LoadingListItemCreator() {
         @Override
         public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-            View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.loading_row, parent, false);
+            View view = LayoutInflater.from(parent.getContext())
+                .inflate(R.layout.loading_row, parent, false);
             return new RecyclerView.ViewHolder(view) {
             };
         }
